recycler view

+++ app/src/main/java/ru/netology/nmedia/dto/Post.kt
@@ -1,12 +1,12 @@
 package ru.netology.nmedia.dto
 
 data class Post(
-    val id: Int,
+    val id: Long,
     val author: String,
     val published: String,
     val content: String,
     var likedByMe: Boolean,
-    var likes: Int,
-    var shares: Int,
+    var likes: Long,
+    var shares: Long,
     var shareByMe: Boolean
 )
